+++ 16_Finish/index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <title>Проект "Комменты"</title>
+  <meta charset="utf-8" />
+  <link rel="stylesheet" href="style.css" />
+</head>
+
+<body>
+  <div class="container" id="app">
+  </div>
+  <script type="module" src="main.js"></script>
+</body>
+
+</html>
