+++ index.html
@@ -2,15 +2,15 @@
 <html>
 
 <head>
-  <title>Проект "Комменты"</title>
   <meta charset="utf-8" />
-  <link rel="stylesheet" href="style.css" />
+  <title>Урок "Асинхронность"</title>
 </head>
 
 <body>
-  <div class="container" id="app">
-  </div>
-  <script type="module" src="main.js"></script>
+  <div id="app"></div>
 </body>
+
+<!-- <script type="module" src="main.js"></script> -->
+<script src="./dist/main.js"></script>
 
 </html>
