+++ main.js
@@ -1,193 +1,155 @@
-import { getTodo, postTodo, toggleLike } from "./api.js";
-import { renderComment } from './render.js';
-import { setToken, token, userName } from './autorize.js';
+// План
+// 1. Реализовать форму логина в приложении (+)
+//  * Перенести всю разметку в рендер функцию (+)
+//  * Сделать форму входа динамическй (+)
+//  * отрефакторить приложение на модули
+//    * api (+)
+//    * вытищть логин компонент в отдельный модуль (+)
+//    * TODO: вытащить компонент списка задач и форму добавления в отдельный модуль
+// 2. Реализовать форму регистрации (+)
 
-const buttonElement = document.getElementById("add-button");
+import { addTodo, deleteTodo, getTodos } from "./api.js";
+import { renderLoginComponent } from "./components/login-component.js";
+import { format } from "date-fns";
+import _ from 'lodash';
+// import { formatDateToRu, formatDateToUs } from "./lib/formatDate/formatDate.js";
+let tasks = [];
 
-export let comments = [];
+let token = "Bearer asb4c4boc86gasb4c4boc86g37k3bk3cg3c03ck3k37w3cc3bo3b8";
 
-function correctDate(date) {
-  let currentDate = new Date(date);
-  let todayDay = currentDate.getDate();
-  let todayMonth = currentDate.getMonth() + 1;
-  let todayYear = String(currentDate.getFullYear()).slice(-2);
-  let todayHours = currentDate.getHours();
-  let todayMinutes = currentDate.getMinutes();
-  todayDay = todayDay < 10 ? "0" + todayDay : todayDay;
-  todayMonth = todayMonth < 10 ? "0" + todayMonth : todayMonth;
-  todayHours = todayHours < 10 ? "0" + todayHours : todayHours;
-  todayMinutes = todayMinutes < 10 ? "0" + todayMinutes : todayMinutes;
+token = null;
 
-  let formattedDate = `${todayDay}.${todayMonth}.${todayYear} ${todayHours}:${todayMinutes} `;
-  return formattedDate;
-}
-
-export function apiFormHide() {//Порячу форму добавления коментария 
-  //console.log("apiFormHide");
-  const apiFormHide = document.querySelector(".add-form");
-  apiFormHide.classList.add("hidden");
-}
-
-export function apiFormShow() {//Показываю форму добавления коментария
-
-  const apiFormShow = document.querySelector(".add-form");
-  apiFormShow.classList.remove("hidden");
-}
-
-function showCommentLoading() {//Показываю (удаляю в стилях блок) временный текст до загрузки комментариев
-
-  const showCommentLoading = document.querySelector(".api-loader");
-  showCommentLoading.classList.remove("hidden");
-}
-
-function hideCommentLoading() {//Удаляю (возвращаю блок) временный текст до загрузки комментариев
-  const hideCommentLoading = document.querySelector(".api-loader");
-  hideCommentLoading.classList.add("hidden");
-}
-
-function showCommentAdd() {
-  const showCommentAdd = document.querySelector(".comment-add")
-  showCommentAdd.classList.remove("hidden")
-}
-
-function hideCommentAdd() {
-  const hideCommentAdd = document.querySelector(".comment-add")
-  hideCommentAdd.classList.add("hidden")
-}
-
-export function showInternetError() {
-  const showInternetError = document.querySelector(".internet-error")
-  showInternetError.classList.remove("hidden")
-}
-export function hideInternetError() {
-  const hideInternetError = document.querySelector(".internet-error")
-  hideInternetError.classList.add("hidden")
-}
-
-export function hideAutorizeForm() {
-  const hideCommentLoading = document.querySelector(".autorizeForm");
-  hideCommentLoading.classList.add("hidden");
-}
-
-export function apiGet() {
-  renderComment();
-  showCommentLoading();
-  getTodo().then((responseData) => {
-    console.log(responseData);
-
-    const apiComment = responseData.comments.map((comment) => {
-      return {
-        id: comment.id,
-        name: comment.author.name,
-        data: correctDate(comment.date),
-        text: comment.text,
-        like: comment.likes,
-        isLiked: comment.isLiked,
-      }
-    });
-    comments = apiComment;
-    hideCommentLoading();
-    renderComment();
+const fetchTodosAndRender = () => {
+  return getTodos({ token }).then((responseData) => {
+    tasks = responseData.todos;
+    renderApp();
   });
-}
+};
 
-apiGet();
+const renderApp = () => {
+  const appEl = document.getElementById("app");
+  if (!token) {
+    renderLoginComponent({
+      appEl,
+      setToken: (newToken) => {
+        token = newToken;
+      },
+      fetchTodosAndRender,
+    });
 
-export function addComment() {
+    return;
+  }
+
+  // const formatDate = (date) => {
+  //   return `${date.getDate() < 10 ? "0" + date.getDate() : date.getDate()}/${
+  //     date.getMonth() < 10 ? "0" + date.getMonth() : date.getMonth()
+  //   }/${date.getFullYear()} ${
+  //     date.getHours() < 10 ? "0" + date.getHours() : date.getHours()
+  //   }:${date.getMinutes() < 10 ? "0" + date.getMinutes() : date.getMinutes()}`;
+  // };
+
+  const country = "ru";
+  const tasksHtml = tasks
+    .map((task) => {
+      // Вызываем функцию format из date-fns, первый параметр — это дата, которую
+      // хотим отформатировать, второй параметр — это строка: к какому формату
+      // желаем привести дату. Обратите внимание MM — это номер месяца,
+      // mm — это минуты
+      //в разметке ЗАДАЧА СОЗДАНА был тернарный оператор в зависимости от переменной формата даты. Мы заменили все на функцию из модуляdate-fns
+      const createDate = format(new Date(task.created_at), 'MM-dd-yyyy hh:mm');
+      return `
+          <li class="task">
+            <p class="task-text">
+              ${task.text} (Создал: ${task.user?.name ?? "Неизвестно"})
+              <button data-id="${task.id
+        }" class="button delete-button">Удалить</button>
+            </p>
+            <p><i>Задача создана123: ${createDate}
+                </i></p>
+          </li>`;
+    })
+    .join("");
+
+  const appHtml = `
+                <h1>Список задач</h1>
+
+                <ul class="tasks" id="list">
+                <!-- Список рендерится из JS -->
+                ${tasksHtml}
+                </ul>
+                <br />
+                <div class="form">
+                <h3 class="form-title">Форма добавления</h3>
+                <div class="form-row">
+                    Что нужно сделать:
+                    <input
+                    type="text"
+                    id="text-input"
+                    class="input"
+                    placeholder="Выпить кофе"
+                    />
+                </div>
+                <br />
+                <button class="button" id="add-button">Добавить</button>
+                </div>`;
+
+  appEl.innerHTML = appHtml;
+
   const buttonElement = document.getElementById("add-button");
-  const nameInputElemnt = document.getElementById("input-name");
-  const commentInputElement = document.getElementById("input-comment");
-  const buttonExit = document.getElementById("exit-button");
+  const listElement = document.getElementById("list");
+  const textInputElement = document.getElementById("text-input");
 
-  buttonExit.addEventListener('click', () => {
-    // alert("До свидания. Возвращайтесь.");
+  const deleteButtons = document.querySelectorAll(".delete-button");
 
-    localStorage.removeItem('token');
-    setToken(null);
-    return apiGet(); //перерендерить 
-  })
+  for (const deleteButton of deleteButtons) {
+    deleteButton.addEventListener("click", (event) => {
+      event.stopPropagation();
 
-  buttonElement.addEventListener('click', () => {
-    nameInputElemnt.classList.remove("error");
-    commentInputElement.classList.remove("error");
-    if (nameInputElemnt.value === "" || commentInputElement.value === "") {
-      alert("Заполните ИМЯ и ваш комментарий, пожалуйста.")
+      const id = deleteButton.dataset.id;
+
+      // подписываемся на успешное завершение запроса с помощью then
+      deleteTodo({
+        id,
+        token,
+      }).then((responseData) => {
+        // получили данные и рендерим их в приложении
+        tasks = responseData.todos;
+        renderApp();
+      });
+    });
+  }
+
+  buttonElement.addEventListener("click", () => {
+    if (textInputElement.value === "") {
       return;
     }
 
-    hideInternetError();
-    showCommentAdd();
+    buttonElement.disabled = true;
+    buttonElement.textContent = "Задача добавляеятся...";
 
-    postTodo({ textInApi: commentInputElement.value, nameInApi: nameInputElemnt.value, token })
-      .then((response) => {
-
-        hideCommentAdd();
-        if (response.status === 500) {
-          throw new Error('500')
-        }
-        if (response.status === 400) {
-          throw new Error('400')
-        }
-
-        apiFormShow();
-        nameInputElemnt.value = "";
-        commentInputElement.value = "";
-        apiGet();
-        hideCommentLoading();
-
+    addTodo({
+      text: _.capitalize(textInputElement.value),
+      // textInputElement.value,
+      token,
+    })
+      .then(() => {
+        // TODO: кинуть исключение
+        textInputElement.value = "";
+      })
+      .then(() => {
+        return fetchTodosAndRender();
+      })
+      .then(() => {
+        buttonElement.disabled = false;
+        buttonElement.textContent = "Добавить";
       })
       .catch((error) => {
-        if (error.message === '500') {
-          alert("Сервер сломаааался.")
-        }
-        if (error.message === '400') {
-          alert("Имя и комментарий должны быть не короче 3 символов.")
-        }
-        if (error.message === 'Failed to fetch') {
-          showInternetError();
-          alert("Кажется, у вас сломался интернет, попробуйте позже.");
-        }
-        console.warn(error);
-      })
-  })
-}
-
-//Написал функцию. которая инициализирует обработчики кликов по кнопкам Лайк
-///////НУЖНО ПЕРЕДАТЬ В РЕНДЕР МОДУЛЬ В КАЧЕСТВЕ ПАРАМЕТРА
-
-export function addLikeEventListeners() {
-  const likeButtons = document.querySelectorAll(".like-button");
-
-  likeButtons.forEach((likeButton, index) => {
-    likeButton.addEventListener("click", (event) => {
-      event.stopPropagation();
-      toggleLike({ id: comments[index].id }).then(() => {
-        apiGet();
-      })
-    });
+        console.error(error);
+        alert("Кажется у вас проблемы с интернетом, попробуйте позже");
+        buttonElement.disabled = false;
+        buttonElement.textContent = "Добавить";
+      });
   });
-}
+};
 
-addLikeEventListeners();
-
-// ///////НУЖНО ПЕРЕДАТЬ В РЕНДЕР МОДУЛЬ В КАЧЕСТВЕ ПАРАМЕТРА
-export function oncommentClickEventListener() {
-
-  const commentUpdate = document.querySelectorAll('.comment');
-  for (const comment of commentUpdate) {
-    comment.addEventListener('click', () => {
-
-      const commentInputElement = document.getElementById("input-comment");
-      console.log("Нажал на коментарий");
-      let index = comment.dataset.index;
-      let object = comments[index];
-      if (userName === comments[index].name) {
-        return;
-      }
-      commentInputElement.value = `${object.text} // ${object.name}`;
-      console.log(commentInputElement.value);
-
-    });
-  }
-}
-
+renderApp();
