+++ package.json
@@ -1,0 +1,31 @@
+{
+    "name": "bl2-hw17",
+    "version": "1.0.0",
+    "description": "Program spisok zadach",
+    "main": "index.js",
+    "directories": {
+        "lib": "lib"
+    },
+    "scripts": {
+        "dev": "webpack --mode development --watch",
+        "start": "http-server -p 3004",
+        "format": "prettier --write ./src/",
+        "lint:js": "eslint ./src/**/*.js"
+    },
+    "author": "Katkov A/",
+    "license": "ISC",
+    "dependencies": {
+        "date-fns": "^2.30.0",
+        "http-server": "^14.1.1",
+        "lodash": "^4.17.21"
+    },
+    "devDependencies": {
+        "axios": "^1.6.2",
+        "eslint": "^8.54.0",
+        "eslint-config-prettier": "^9.0.0",
+        "eslint-plugin-prettier": "^5.0.1",
+        "prettier": "^3.1.0",
+        "webpack": "^5.89.0",
+        "webpack-cli": "^5.1.4"
+    }
+}
