+++ src/index.js
@@ -1,0 +1,5 @@
+const add = (a, b) => {
+    return a + b + ""
+}
+
+add(1, 2)
