+++ webpack.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+    entry: "./main.js", // Входной файл, в котором мы пишем свой код
+    output: {
+        filename: "main.js" // Выходной файл, который подключаем к HTML
+        // Обратите внимание, сохранится он по пути "./dist/main.js"
+    }
+}
